Moving smith_wilson functions to separate module
